requirements.txt
streamlit
pandas
plotly
google-auth
google-api-python-client
openpyxl
numpy
